[IMP] vault: Update value when input is decrypted
Up until now, when a field was decrypted, making a change would cause
it to be lost, which could confuse the user. After fixing this, it
happened that discarding the changes made it seem like the change had
not been discarded.

The measure that has been taken is to, upon discarding, do as saving
does: hide the value of the field, and the change remains in the input
as long as it is not saved, discarded, or hidden.